add basic logger instance
