+++ Sources/ToucanSDK/Extensions/FileManager+Extensions.swift
@@ -9,6 +9,32 @@ import Foundation
 
 extension FileManager {
 
+    func copyRecursively(
+        from inputURL: URL,
+        to outputURL: URL
+    ) throws {
+        guard directoryExists(at: inputURL) else {
+            return
+        }
+        if !directoryExists(at: outputURL) {
+            try createDirectory(at: outputURL)
+        }
+
+        for item in listDirectory(at: inputURL) {
+            let itemSourceUrl = inputURL.appendingPathComponent(item)
+            let itemDestinationUrl = outputURL.appendingPathComponent(item)
+            if fileExists(at: itemSourceUrl) {
+                if fileExists(at: itemDestinationUrl) {
+                    try delete(at: itemDestinationUrl)
+                }
+                try copy(from: itemSourceUrl, to: itemDestinationUrl)
+            }
+            else {
+                try copyRecursively(from: itemSourceUrl, to: itemDestinationUrl)
+            }
+        }
+    }
+
     func createParentFolderIfNeeded(for url: URL) throws {
         let folderPath =
             "/"
