+++ Sources/toucan-cli/Commands/Watch.swift
@@ -1,6 +1,7 @@
 import Foundation
 import ArgumentParser
 import ToucanSDK
+import Logging
 
 #if os(macOS)
 import Dispatch
@@ -28,12 +29,20 @@ extension Entrypoint {
 
         @Option(name: .shortAndLong, help: "The base url to use.")
         var baseUrl: String? = nil
+        
+        @Option(name: .shortAndLong, help: "The log level to use.")
+        var logLevel: Logger.Level = .info
+
 
         mutating func run() async throws {
+            var logger = Logger(label: "toucan")
+            logger.logLevel = logLevel
+            
             let toucan = Toucan(
                 input: input,
                 output: output,
-                baseUrl: baseUrl
+                baseUrl: baseUrl,
+                logger: logger
             )
             try toucan.generate()
 
@@ -55,26 +64,26 @@ extension Entrypoint {
                         return
                     }
 
-                    print("Generating site...")
+                    logger.info("Generating site...")
                     do {
                         try toucan.generate()
                         lastGenerationTime = now
                     }
                     catch {
-                        print("\(error)")
+                        logger.error("\(error)")
                     }
-                    print("Site re-generated.")
+                    logger.info("Site re-generated.")
                 }
             )
 
             eventStream.setDispatchQueue(DispatchQueue.main)
 
             try eventStream.start()
-            print("👀 Watching: `\(input)` -> \(output).")
+            logger.info("👀 Watching: `\(input)` -> \(output).")
 
             waitForever()
             #else
-            print("👀 This is a macOS only feature for now.")
+            logger.info("👀 This is a macOS only feature for now.")
             #endif
         }
     }
