+++ Sources/toucan-cli/Extensions/Logger+Extensions.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  toucan
+//
+//  Created by Tibor Bodecs on 2024. 10. 06..
+//
+
+import Logging
+import ArgumentParser
+
+extension Logger.Level: @retroactive ExpressibleByArgument {
+    
+}
